requirements.txt
python-bitcoinrpc==1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9 
